Implemented Difficulty Properly
Added complex difficulty functionality.
The difficulty value changes during gameplay based on time played, amount of enemies hit in a row and enemies "fouled".
Size and movement speed of the Player and Enemies is scaled to difficulty, meaning the game zooms out and the play field increases in size.
The Spawner now has a list of when and how enemies can spawn, based on difficulties ranges and some other factors, and I added an initial "level map" of what enemies to spawn when.
Also added a few stats for various things